Bump Acontplus.Core to version 1.2.5 and refactor error handling extensions
- Updated project version to 1.2.5 in Acontplus.Core.csproj.
- Removed `DomainErrorsExtensions` to streamline code structure.
- Enhanced `DomainErrorExtensions` with refined methods for consistent and efficient API response handling.
- Simplified error severity logic and improved method organization for better maintainability.
